requirements.txt
Flask==3.0.3

firebase-admin

python-dotenv